+++ aksel.nav.no/website/sanity/config.ts
@@ -57,7 +57,6 @@ export const templatesKategorier = [
   { title: "Brev", value: "brev" },
   { title: "Støtte", value: "stotte" },
   { title: "Søknadsdialog", value: "soknadsdialog" },
-  { title: "Skjemavalidering", value: "skjemavalidering" },
 ];
 
 export const bloggKategorier = [