+++ examples/referansesider/src/main.tsx
@@ -1,0 +1,27 @@
+import { RouterProvider, createRouter } from "@tanstack/react-router";
+import { StrictMode } from "react";
+import ReactDOM from "react-dom/client";
+import "@navikt/ds-tokens/dist/darkside/tokens.css";
+import "./index.css";
+import { routeTree } from "./routeTree.gen";
+import { ThemeProvider } from "./theme/ThemeProvider";
+
+const router = createRouter({ routeTree, trailingSlash: "always" });
+
+declare module "@tanstack/react-router" {
+  interface Register {
+    router: typeof router;
+  }
+}
+
+const rootElement = document.getElementById("root")!;
+if (!rootElement.innerHTML) {
+  const root = ReactDOM.createRoot(rootElement);
+  root.render(
+    <StrictMode>
+      <ThemeProvider>
+        <RouterProvider router={router} />
+      </ThemeProvider>
+    </StrictMode>,
+  );
+}