+++ @navikt/aksel-icons/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @navikt/aksel-icons
 
+## 7.1.2
+
+## 7.1.1
+
+## 7.1.0
+
+## 7.0.1
+
+## 7.0.0
+
+### Major Changes
+
+- Icons: Removed renamed icons. [See migration guide for changes](https://aksel.nav.no/grunnleggende/kode/migrering#194b60833d9e). ([#3139](https://github.com/navikt/aksel/pull/3139))
+
+## 6.17.0
+
+### Minor Changes
+
+- Icons: New 'face' icons for 'Engagement'-category. ([#3134](https://github.com/navikt/aksel/pull/3134))
+
 ## 6.16.3
 
 ### Patch Changes