+++ @navikt/core/tailwind/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @navikt/ds-tailwind
 
+## 7.1.2
+
+## 7.1.1
+
+## 7.1.0
+
+## 7.0.1
+
+## 7.0.0
+
+### Major Changes
+
+- Tailwind: Extended 'screens'-config in theme to match Aksel breakpoints. Tailwind and Primitives can now be used side by side with matching dynamic breakpoints. [See migration guide for potential issues when adopting](https://aksel.nav.no/grunnleggende/kode/migrering#3a2340f6f69b). ([#3139](https://github.com/navikt/aksel/pull/3139))
+
+## 6.17.0
+
 ## 6.16.3
 
 ## 6.16.2